Fetch country description from Wikipedia API
Created a function in scripts/fetchDataSets.js for fetching the Wikipedia API. It has to do it two times, because we first need to find what the exact title for the page is before we can look up that page.
The function returns the HTML of the description that wikipedia gives back.